Summarize metadata processing handlers to one class
